+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fixed
 - Subset with BBOX does not modify the input model anymore (#10)
+- `cityjson.load()` does not fail on a `GeometryInstance`, however it does not load it either (#19)
 
 
 ## [0.7.0] - 2021-12-01